backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
google-generativeai
pillow